refactor: Remove the hyphen from the name
