+++ pkg/crd/client/policyreport/clientset/versioned/scheme/register.go
