+++ pkg/crd/client/policyreport/clientset/versioned/typed/policyreport/v1alpha2/clusterpolicyreport.go
@@ -28,7 +28,7 @@ import (
 	rest "k8s.io/client-go/rest"
 
 	v1alpha2 "github.com/kyverno/policy-reporter/pkg/crd/api/policyreport/v1alpha2"
-	scheme "github.com/kyverno/policy-reporter/pkg/crd/client/policy-report/clientset/versioned/scheme"
+	scheme "github.com/kyverno/policy-reporter/pkg/crd/client/policyreport/clientset/versioned/scheme"
 )
 
 // ClusterPolicyReportsGetter has a method to return a ClusterPolicyReportInterface.
