+++ pkg/crd/client/policyreport/clientset/versioned/typed/policyreport/v1alpha2/fake/fake_policyreport.go
