+++ pkg/crd/client/policyreport/clientset/versioned/typed/policyreport/v1alpha2/policyreport_client.go
@@ -24,7 +24,7 @@ import (
 	rest "k8s.io/client-go/rest"
 
 	v1alpha2 "github.com/kyverno/policy-reporter/pkg/crd/api/policyreport/v1alpha2"
-	"github.com/kyverno/policy-reporter/pkg/crd/client/policy-report/clientset/versioned/scheme"
+	"github.com/kyverno/policy-reporter/pkg/crd/client/policyreport/clientset/versioned/scheme"
 )
 
 type Wgpolicyk8sV1alpha2Interface interface {
