+++ pkg/crd/client/policyreport/informers/externalversions/policyreport/v1alpha2/interface.go
@@ -19,7 +19,7 @@ limitations under the License.
 package v1alpha2
 
 import (
-	internalinterfaces "github.com/kyverno/policy-reporter/pkg/crd/client/policy-report/informers/externalversions/internalinterfaces"
+	internalinterfaces "github.com/kyverno/policy-reporter/pkg/crd/client/policyreport/informers/externalversions/internalinterfaces"
 )
 
 // Interface provides access to all the informers in this group version.
