+++ pkg/crd/client/policyreport/listers/policyreport/v1alpha2/clusterpolicyreport.go
