+++ pkg/email/violations/fixtures_test.go
@@ -3,8 +3,8 @@ package violations_test
 import (
 	"go.uber.org/zap"
 
-	"github.com/kyverno/policy-reporter/pkg/crd/client/policy-report/clientset/versioned/fake"
-	v1alpha2client "github.com/kyverno/policy-reporter/pkg/crd/client/policy-report/clientset/versioned/typed/policyreport/v1alpha2"
+	"github.com/kyverno/policy-reporter/pkg/crd/client/policyreport/clientset/versioned/fake"
+	v1alpha2client "github.com/kyverno/policy-reporter/pkg/crd/client/policyreport/clientset/versioned/typed/policyreport/v1alpha2"
 	"github.com/kyverno/policy-reporter/pkg/email"
 	"github.com/kyverno/policy-reporter/pkg/validate"
 )
