+++ pkg/kubernetes/fixtures_test.go
@@ -7,8 +7,8 @@ import (
 	metafake "k8s.io/client-go/metadata/fake"
 
 	pr "github.com/kyverno/policy-reporter/pkg/crd/api/policyreport/v1alpha2"
-	"github.com/kyverno/policy-reporter/pkg/crd/client/policy-report/clientset/versioned/fake"
-	v1alpha2client "github.com/kyverno/policy-reporter/pkg/crd/client/policy-report/clientset/versioned/typed/policyreport/v1alpha2"
+	"github.com/kyverno/policy-reporter/pkg/crd/client/policyreport/clientset/versioned/fake"
+	v1alpha2client "github.com/kyverno/policy-reporter/pkg/crd/client/policyreport/clientset/versioned/typed/policyreport/v1alpha2"
 	"github.com/kyverno/policy-reporter/pkg/report"
 )
